improvements

+++ PerhapsEngineNative/Engine/Core/Systems/EventSystem/EventSystem.cpp
@@ -3,6 +3,6 @@
 namespace Perhaps
 {
 	std::map<Event::DescriptorType, std::vector<EventDispatcher::StoredFunction*>> EventDispatcher::observers;
-	std::map<int, EventDispatcher::StoredFunction*> EventDispatcher::idToFun;
-	int EventDispatcher::idReel;
+	std::map<size_t, EventDispatcher::StoredFunction*> EventDispatcher::idToFun;
+	size_t EventDispatcher::idReel;
 }
